Moved Heatmap structs to seperate files
