+++ AirCasting/Map/Heatmap/Heatmap.swift
@@ -185,148 +185,4 @@ struct Heatmap {
     private func gridSquareKey(_ x: Int, _ y: Int) -> String {
         "\(x)_\(y)"
     }
-    
-    struct GridCalculator {
-        let x: Int
-        let y: Int
-        let visibleRegion: VisibleRegionCoordinates
-        let lonGridSize: Double
-        let latGridSize: Double
-        
-        init(_ x: Int, _ y: Int, _ visibleRegion: VisibleRegionCoordinates, _ lonGridSize: Double, _ latGridSize: Double) {
-            self.x = x
-            self.y = y
-            self.visibleRegion = visibleRegion
-            self.lonGridSize = lonGridSize
-            self.latGridSize = latGridSize
-        }
-        
-        func southWestLatLng() -> CLLocationCoordinate2D {
-            CLLocationCoordinate2D(latitude: southLatitude(), longitude: westLongitude())
-        }
-        
-        func southEastLatLng() -> CLLocationCoordinate2D {
-            CLLocationCoordinate2D(latitude: southLatitude(), longitude: eastLongitude())
-        }
-        
-        func northEastLatLng() -> CLLocationCoordinate2D {
-            CLLocationCoordinate2D(latitude: northLatitude(), longitude: eastLongitude())
-        }
-        
-        func northWestLatLng() -> CLLocationCoordinate2D {
-            CLLocationCoordinate2D(latitude: northLatitude(), longitude: westLongitude())
-        }
-        
-        private func eastLongitude() -> Double {
-            visibleRegion.lonWest + Double(x) * lonGridSize
-        }
-        
-        private func westLongitude() -> Double {
-            visibleRegion.lonWest + Double(x - 1) * lonGridSize
-        }
-        
-        private func northLatitude() -> Double {
-            visibleRegion.latSouth + Double(y) * latGridSize
-        }
-        
-        private func southLatitude() -> Double {
-            visibleRegion.latSouth + Double(y - 1) * latGridSize
-        }
-    }
-    
-    struct GridSquare {
-        private var mapView: GMSMapView! = nil
-        private var sum: Double = 0.0
-        private var number: Int = 0
-        private var averagedValue: Double? = nil
-        private var fillColor: UIColor? = nil
-        private var sensorThreshold: SensorThreshold! = nil
-        
-        private var polygon: GMSPolygon? = nil
-        private var newColor: Bool = false
-        
-        private var southWestLatLng: CLLocationCoordinate2D
-        private var southEastLatLng: CLLocationCoordinate2D
-        var northEastLatLng: CLLocationCoordinate2D
-        private var northWestLatLng: CLLocationCoordinate2D
-        
-        init(mapView: GMSMapView, sensorThreshold: SensorThreshold, _ southWestLatLng: CLLocationCoordinate2D, _ southEastLatLng: CLLocationCoordinate2D, _ northEastLatLng: CLLocationCoordinate2D, _ northWestLatLng: CLLocationCoordinate2D) {
-            self.mapView = mapView
-            self.sensorThreshold = sensorThreshold
-            self.southWestLatLng = southWestLatLng
-            self.southEastLatLng = southEastLatLng
-            self.northEastLatLng = northEastLatLng
-            self.northWestLatLng = northWestLatLng
-            
-            let rect = GMSMutablePath()
-            rect.add(southWestLatLng)
-            rect.add(southEastLatLng)
-            rect.add(northEastLatLng)
-            rect.add(northWestLatLng)
-            self.polygon = GMSPolygon(path: rect)
-        }
-        
-        func inBounds(coordinates: CLLocationCoordinate2D?) -> Bool {
-            let bounds = GMSCoordinateBounds(coordinate: southWestLatLng, coordinate: northEastLatLng)
-            
-            return coordinates != nil ? bounds.contains(coordinates!) : false
-        }
-        
-        mutating func addMeasurement(_ pathPoint: PathPoint) {
-            sum += pathPoint.measurement
-            number += 1
-            calculateAverage()
-            
-            guard let averagedValue = averagedValue else { return }
-            let color: UIColor = GoogleMapView.color(value: Int32(averagedValue), threshold: sensorThreshold).withAlphaComponent(0.5)
-            if color != fillColor {
-                fillColor = color
-                newColor = true
-                polygon?.fillColor = color
-            }
-        }
-        
-        mutating func drawPolygon() {
-            if (shouldDrawPolygon()) {
-                addPolygon()
-            }
-        }
-        
-        mutating func addPolygon() {
-            if (polygon?.map != nil) {
-                polygon?.map = nil
-            }
-            
-            polygon?.map = mapView
-            newColor = false
-        }
-        
-        func shouldDrawPolygon() -> Bool {
-            return self.newColor || (polygon?.map == nil && averagedValue != nil)
-        }
-        
-        
-        mutating func calculateAverage() {
-            self.averagedValue = sum / Double(number)
-        }
-        
-        func remove() {
-            polygon?.map = nil
-        }
-    }
-    
-    struct VisibleRegionCoordinates {
-        var latNorth: CLLocationDegrees
-        var latSouth: CLLocationDegrees
-        var lonEast: CLLocationDegrees
-        var lonWest: CLLocationDegrees
-        
-        init(_ latNorth: CLLocationDegrees, _ latSouth: CLLocationDegrees, _ lonEast: CLLocationDegrees, _ lonWest: CLLocationDegrees) {
-            self.latNorth = latNorth
-            self.latSouth = latSouth
-            self.lonEast = lonEast
-            self.lonWest = lonWest
-        }
-        
-    }
 }
